refactor: move BboxOverlay to shared component
